ref(ts): Refactor Incident Details: Related Issues to typescript
Just typescript things
